requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
biopython>=1.81